+++ changelog.md
@@ -1,6 +1,14 @@
 # Changelog of score2dx
 
-* v1.0.0 [2021-10-08]:
+* 2.0.0 [2021-10-21]:
+    * Update for IIDX29 CastHour.
+    * Json MusicDatabase change to v2 format.
+        * Update Json MusicDatabase to include IIDX29 (42 musics).
+        * Fix many incorrect info in Json MusicDatabase.
+    * Refactor C++ interface.
+        * In C++ `MusicDatabase` is also class now, separated from `Core`.
+
+* 1.0.0 [2021-10-08]:
     * Initial commit with basic functionalities.
         * Load CSV.
         * Import/Export.